server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
finnhub-python==2.4.18
Flask==2.3.2
frozendict==2.3.8
frozenlist==1.4.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.25.2
openai==0.27.8
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.7
yarl==1.9.2
yfinance==0.2.28
